pyproject.toml
[tool.poetry]
name = "soft-eng-challenge"
version = "0.1.0"
description = "Assesment of flugel.it"
authors = ["Omar Faruk Riyad <omar@mixrank.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "3.2.12"
django-rest-framework = "^0.1.0"
names = "^0.3.0"
confidential = "^2.3.1"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"
pytest-xdist = "^2.5.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
